Factor out `run_deploy_jar.py`, to simplify rule graphs and dependencies
